+++ src/components/Header.tsx
@@ -17,6 +17,7 @@ import { useForm } from 'react-hook-form';
 import { getAllUsers } from 'api/usersServices';
 import { addNewBoard, getAllBoardsOfServer } from 'api/boardsService';
 import { IAddBoardForm, IBoardsOfUser } from 'types/types';
+import Navigation from '../constants/navigation';
 
 function Header() {
   const token = useAppSelector((state) => state.auth.token);
